feat: question serviceImpl 관련 오답문제 수정

+++ server/src/main/java/com/ssafy/odab/domain/question_result/service/QuestionResultService.java
@@ -15,5 +15,5 @@ public interface QuestionResultService {
 
     WrongQuestionResponseDto findRecentWrongAnswersBySchoolLevel(String schoolLevel, Integer userId, LocalDateTime startTime);
 
-    WrongQuestionResponseDto findRecentWrongAnswersByGrade(Integer grade, Integer userId, LocalDateTime startTime);
+    WrongQuestionResponseDto findRecentWrongAnswersByGrade(Byte grade, Integer userId, LocalDateTime startTime);
 }
